% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSpectraModel.R
\docType{methods}
\name{lnLikelihood}
\alias{lnLikelihood}
\alias{likelihood-FuncModel}
\alias{criteria}
\alias{criteria-FuncModel}
\title{get parts of a class [\code{FuncModel}]

get log-likelihoods from a [\code{FuncModel}] S3 object.}
\usage{
lnLikelihood(x)

criteria(x)
}
\arguments{
\item{x}{an object of class [\code{FuncModel}]}
}
\examples{
## the famous formosat data set
\dontrun{
  d <-readFiles(firstYear=2012, lastYear=2012)
  r <- learnFuncSpectra(d)
  likelihood(r)
  }

## the famous formosat data set
\dontrun{
  d <-readFiles(firstYear=2012, lastYear=2012)
  r <- learnFuncSpectra(d)
  criteria(r)
  }

}
