% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSpectraModel.R
\docType{class}
\name{FuncSpectraModel}
\alias{FuncSpectraModel}
\alias{FuncSpectraModel-class}
\title{Definition of the [\code{\linkS4class{FuncSpectraModel}}] class}
\description{
This class defines a Gaussian model for the spectrum with mean
and variance varying along the time. It inherits
from [\code{\linkS4class{ICloHeModel}}].
}
\section{Slots}{

\describe{
\item{\code{mu}}{matrix with the sampled (100 times) mean spectrum.}

\item{\code{sigma2}}{vector with the variance of each spectrum}
}}

\examples{
getSlots("FuncSpectraModel")

}
\seealso{
[\code{\linkS4class{ICloHeModel}}] class
}
\author{
Serge Iovleff
}
