% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianMutSigmatModel.R
\docType{class}
\name{GaussianMutSigmatModel}
\alias{GaussianMutSigmatModel}
\alias{GaussianMutSigmatModel-class}
\title{Definition of the [\code{\linkS4class{GaussianMutSigmatModel}}] class}
\description{
This class defines a Gaussian model for the spectrum with mean
and variance varying along the time. It inherits
from [\code{\linkS4class{ICloHeModel}}].
}
\section{Slots}{

\describe{
\item{\code{mut}}{A list with the vectors of the means.}

\item{\code{sigmat}}{A list with the variances matrices.}
}}

\examples{
getSlots("GaussianMutSigmatModel")

}
\seealso{
[\code{\linkS4class{ICloHeModel}}] class
}
\author{
Serge Iovleff
}
