% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICloHeModel.R
\docType{class}
\name{ICloHeModel}
\alias{ICloHeModel}
\alias{ICloHeModel-class}
\title{Interface base Class [\code{\linkS4class{ICloHeModel}}] for CloHe models.}
\description{
This class encapsulate the common parameters of all the CloHe models.
}
\section{Slots}{

\describe{
\item{\code{classNumber}}{the integer with the class attributed to the pixel}

\item{\code{classLabel}}{the integer attributed to the pixel}

\item{\code{nk}}{the number of sample in this class}

\item{\code{pk}}{the proportion of sample in this class}

\item{\code{nbSpectrum}}{number of spectrum to analyze}

\item{\code{lnLikelihood}}{log-likelihood of this}

\item{\code{criterion}}{the value of the criterion}

\item{\code{nbFreeParameter}}{number of free parameter}
}}

\examples{
  getSlots("ICloHeModel")

}
\author{
Serge Iovleff
}
