% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICloHeModel.R, R/GaussianMutSigmatModel.R,
%   R/FuncSpectraModel.R
\docType{methods}
\name{initialize,ICloHeModel-method}
\alias{initialize,ICloHeModel-method}
\alias{initialize,GaussianMutSigmatModel-method}
\alias{initialize,FuncSpectraModel-method}
\title{Initialize an instance of a CloHe S4 class.}
\usage{
\S4method{initialize}{ICloHeModel}(.Object)

\S4method{initialize}{GaussianMutSigmatModel}(.Object)

\S4method{initialize}{FuncSpectraModel}(.Object)
}
\description{
Initialization method of the [\code{\linkS4class{ICloHeModel}}] class.
Used internally in the 'CloHe' package.

Initialization method of the [\code{\linkS4class{GaussianMutSigmatModel}}] class.
Used internally in the 'MixAll' package.

Initialization method of the [\code{\linkS4class{FuncSpectraModel}}] class.
Used internally in the 'CloHe' package.
}
\keyword{internal}
