% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSpectraModel.R
\name{learnFuncSpectra}
\alias{learnFuncSpectra}
\title{Create an instance of the [\code{\linkS4class{FuncSpectraModel}}] class}
\usage{
learnFuncSpectra(data, kernelName = "gaussian", width = 50, dim = 5,
  basisName = "Bspline", posKnots = "density", degree = 3,
  criterion = "BIC", maxValue = NULL, test = NULL)
}
\arguments{
\item{data}{list containing the data.}

\item{kernelName}{[\code{string}] with the kernel to use for the covariance matrix.
Available kernels are "gaussian", "exponential", "rationalQuadratic".
Default is "gaussian".}

\item{width}{the width of the kernel to use. Default is 50.}

\item{dim}{[\code{\link{integer}}] number of base to use. Default is 5.}

\item{basisName}{[\code{string}] with the name of the basis to use}

\item{posKnots}{position for the knots of the B-spline. Available position are "periodic", "density",
"uniform". Default is "density".}

\item{degree}{degree of the B-splines. Default is 3.}

\item{criterion}{character defining the criterion to select the best model.
The best model is the one with the lowest criterion value.
Possible values: "BIC", "AIC", "ML". Default is "BIC".}

\item{test}{list with test data to classify}
}
\value{
An instance of the [\code{\linkS4class{FuncModel}}] class.
}
\description{
This function ajust the parameters of a functionnal gaussian curve model.
}
\examples{
## the famous formosat data set
#d <-readSplitFormosatData(path="./data/Formosat/",firstYear=2008, lastYear=2012)
data("multiFormosat")
model <- learnFuncSpectra(multiFormosat)
## get summary
summary(model)

## use graphics functions
\dontrun{
plot(model)
plot(model$models[[1]])
}

## print model
\dontrun{
print(model)
}

}
\author{
Serge Iovleff
}
