% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSpectraModel.R
\docType{methods}
\name{plot.FuncModel}
\alias{plot.FuncModel}
\alias{plot-FuncModel}
\title{Plotting of a class [\code{FuncModel}]}
\usage{
plot.FuncModel(x, ...)
}
\arguments{
\item{x}{an object of class [\code{FuncModel}]}

\item{...}{further arguments passed to or from other methods}

\item{y}{not used. Will be silently ignored}
}
\description{
Plotting data from a [\code{FuncModel}] S3 object
using the estimated parameters.
}
\examples{
## the famous formosat data set
\dontrun{
  d <-readFiles(firstYear=2012, lastYear=2012)
  r <- learnFuncSpectra(d)
  plot(r)
  }

}
\seealso{
\code{\link{plot}}
}
