% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSpectraModel.R
\docType{methods}
\name{plot,FuncSpectraModel,missing-method}
\alias{plot,FuncSpectraModel,missing-method}
\alias{plot-FuncSpectraModel}
\title{Plotting of a class [\code{\linkS4class{FuncSpectraModel}}]}
\usage{
\S4method{plot}{FuncSpectraModel,missing}(x, y, ...)
}
\arguments{
\item{x}{an object of class [\code{\linkS4class{FuncSpectraModel}}]}

\item{...}{further arguments passed to or from other methods}
}
\description{
Plotting data from a [\code{\linkS4class{FuncSpectraModel}}] object
using the estimated parameters.
}
\examples{
 ## the famous formosat data set
\dontrun{
  d <-readFiles(firstYear=2012, lastYear=2012)
  model <- learnFuncSpectra(d)
  plot(r$models[[1]])
  }

}
\seealso{
\code{\link{plot}}
}
