% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSpectraModel.R
\docType{methods}
\name{print.FuncModel}
\alias{print.FuncModel}
\alias{print-FuncModel}
\title{print of a class [\code{FuncModel}]}
\usage{
print.FuncModel(x)
}
\arguments{
\item{x}{an object of class [\code{FuncModel}]}
}
\description{
print data from a [\code{FuncModel}] S3 object.
}
\examples{
## the famous formosat data set
\dontrun{
  d <-readFiles(firstYear=2012, lastYear=2012)
  r <- learnFuncSpectra(d)
  print(r)
  }

}
