% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ICloHeModel.R,
%   R/GaussianMutSigmatModel.R, R/FuncSpectraModel.R
\docType{methods}
\name{print}
\alias{print}
\alias{print,ICloHeModel-method}
\alias{print,GaussianMutSigmatModel-method}
\alias{print}
\alias{GaussianMutSigmatModel-method}
\alias{print,FuncSpectraModel-method}
\alias{print}
\alias{FuncSpectraModel-method}
\title{Print a CloHe S4 class to standard output.}
\usage{
\S4method{print}{ICloHeModel}(x, ...)

\S4method{print}{GaussianMutSigmatModel}(x, ...)

\S4method{print}{FuncSpectraModel}(x, ...)
}
\arguments{
\item{x}{a CloHe object: a \code{\linkS4class{GaussianMutSigmatModel}} or
a \code{\linkS4class{FuncSpectraModel}}.}

\item{...}{further arguments passed to or from other methods}
}
\value{
NULL. Prints to standard out.
}
\description{
Print a CloHe S4 class to standard output.
}
\seealso{
\code{\link{print}}
}
