% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFormosatFiles.R
\name{readFormosatData}
\alias{readFormosatData}
\alias{readMultiFormosatData}
\title{Read the csv formosat files by concatenating all the years in a single series.}
\usage{
readFormosatData(path = "./data/Formosat/", firstYear = 2008,
  lastYear = 2014)

readMultiFormosatData(path = "./data/Formosat/", firstYear = 2008,
  lastYear = 2014)
}
\arguments{
\item{path}{a string with the path to the files}

\item{firstYear}{first year of observations}

\item{lastYear}{last year of observations}

\item{path}{a string with the path to the files}

\item{firstYear}{first year of observations}

\item{lastYear}{last year of observations}
}
\value{
a list with the labels of each observations, the days of each observation (starting from
day 1 of firstYear), the four spectrums (b,g,r,ir), the clouds presence.

a list with the labels of each observations, the days of each observation (starting from
day 1 of firstYear), the four spectrums (b,g,r,ir), the clouds presence.
}
\description{
Each file is in the format i_year.csv, with i=1  for blue spectrum, i=2 for green spectrum
i=3 for red spectrum and i=4 for infra-red spectrum.

Each file is in the format i_year.csv, with i=1  for blue spectrum, i=2 for green spectrum
i=3 for red spectrum and i=4 for infra-red spectrum.
}
