% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ICloHeModel.R,
%   R/GaussianMutSigmatModel.R, R/FuncSpectraModel.R
\docType{methods}
\name{show}
\alias{show}
\alias{show,ICloHeModel-method}
\alias{show}
\alias{show,GaussianMutSigmatModel-method}
\alias{show-GaussianMutSigmatModel,GaussianMutSigmatModel,GaussianMutSigmatModel-method}
\alias{show,FuncSpectraModel-method}
\alias{show-FuncSpectraModel,FuncSpectraModel,FuncSpectraModel-method}
\title{Show description of a CloHe S4 class to standard output.}
\usage{
\S4method{show}{ICloHeModel}(object)

\S4method{show}{GaussianMutSigmatModel}(object)

\S4method{show}{FuncSpectraModel}(object)
}
\arguments{
\item{object}{a MixAll object: a \code{\linkS4class{GaussianMutSigmatModel}},
or a \code{\linkS4class{FuncSpectraModel}}.}
}
\value{
NULL. Prints to standard out.
}
\description{
Show description of a CloHe S4 class to standard output.
}
\seealso{
\code{\link{show}}
}
