% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FuncSpectraModel.R
\docType{methods}
\name{summary.FuncModel}
\alias{summary.FuncModel}
\alias{summary-FuncModel}
\title{summary of a class [\code{FuncModel}]}
\usage{
summary.FuncModel(x)
}
\arguments{
\item{x}{an object of class [\code{FuncModel}]}
}
\description{
summary data from a [\code{FuncModel}] S3 object.
}
\examples{
## the famous formosat data set
\dontrun{
  d <-readFiles(firstYear=2012, lastYear=2012)
  r <- learnFuncSpectra(d)
  summary(r)
  }

}
