% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R, R/ICloHeModel.R,
%   R/GaussianMutSigmatModel.R, R/FuncSpectraModel.R
\docType{methods}
\name{summary}
\alias{summary}
\alias{summary,ICloHeModel-method}
\alias{summary}
\alias{summary,GaussianMutSigmatModel-method}
\alias{summary}
\alias{summary,FuncSpectraModel-method}
\alias{summary}
\title{Produce summary of a CloHe S4 class.}
\usage{
\S4method{summary}{ICloHeModel}(object, ...)

\S4method{summary}{GaussianMutSigmatModel}(object, ...)

\S4method{summary}{FuncSpectraModel}(object, ...)
}
\arguments{
\item{object}{any cluster model deriving from a \code{\linkS4class{ICloheModel}} object.}

\item{...}{further arguments passed to or from other methods}
}
\value{
NULL. Summaries to standard out.
}
\description{
Produce summary of a CloHe S4 class.
}
